dockerfile-runtime-pyspark/requirements.txt
findspark
ibm-watson-machine-learning
pyspark2pmml
wget
